Test fixes (test in wrong file!), delete commented code.
