+++ app/views/events/hospitality_registration_assignments/index.html.erb
@@ -35,7 +35,7 @@
           </div>
         </div>
         <div class="panel-footer">
-          <%= link_to 'Back', event_path(@event, anchor: 'hospitality_info_widget'), class: 'btn btn-default btn-sm' %>
+          <%= link_to 'Back', event_path(@event), class: 'btn btn-default btn-sm' %>
         </div>
     </div>
   </div>
